requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
numpy
opencv-python
moviepy
pyttsx3
mediapipe
fer
insightface
onnxruntime-gpu           
deepface
facenet-pytorch
keras-vggface
keras-applications
realesrgan
